Implement Client Timeout

+++ agriculture_bot_client/src/main/kotlin/Missions.kt
@@ -7,7 +7,7 @@ sealed class Mission() {
     abstract val plant: Plant
     abstract val _id: String
     abstract var resultData: MissionResultData?
-    abstract var processingExpirationDate: Int?
+    abstract var processingExpirationDate: Long?
 }
 
 @Serializable
@@ -16,7 +16,7 @@ data class InspectionMission(
     override val plant: Plant,
     override val _id: String = newId<Mission>().toString(),
     override var resultData: MissionResultData? = null,
-    override var processingExpirationDate: Int? = null,
+    override var processingExpirationDate: Long? = null,
 ) :
     Mission()
 
@@ -27,5 +27,5 @@ data class WateringMission(
     val quantity: Int,
     override val _id: String = newId<Mission>().toString(),
     override var resultData: MissionResultData? = null,
-    override var processingExpirationDate: Int? = null,
+    override var processingExpirationDate: Long? = null,
 ) : Mission()
